+++ src/modules/message/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './send-message.dto';
+export * from './send-interactive-cta.dto';